fix: comprehensive audit fixes for migrations, models, transactions, tasks, permissions, concurrency
- Migrations: alembic.ini (prepend_sys_path, timezone, URL), env.py (lazy db_url, compare_server_default, render_as_batch, literal_binds), script.py.mako (postgresql imports), models (_mixins.py timezone-aware datetime, server_default func.now()), user model (CheckConstraints, server_default, UserCase PK fix), models (BigInteger for snowflakes, CheckConstraints, server_default, indexes, FK ondelete)
- Transactions: uow.py (retry on serialization, isolation level, proper commit/rollback logging), config.py (readonly param for specified_guild_config), count_message (split read/write phases), roulette modal (move heavy calc outside lock), roulette double-commit fix (single TX with message_id), voice activity (early return flag), battlepass claim (atomic), uow retry on serialization, specified_guild_config readonly param
- Tasks: rainbow_role (sleep, batch schedule key by role_id, schedule before discord), delete_ticket (sleep, dispatch after commit), multiplayer_roulette (sleep top loop, lock user rows, single TX), temp_role/temp_multiplier (FOR UPDATE SKIP LOCKED, re-fetch with lock), unpunish (dispatch before delete, per-item try), delete_ticket (dispatch after commit), delete_role_request (continue vs return, edit after commit), temp_multiplier (for_update on re-fetch), clan_reputation (per-item try)
- Events: voice gateway (await gather, ensure_voice_channel_exists, stale state fix), channel update (UnboundLocalError, if new_topic or new_topic), private rooms (unique per guild, ON CONFLICT DO NOTHING)
- Permissions: decorator (guild None check, deny empty list, UNSAFE audit, guild_only wrapper), jwt (Bearer strip, exp/aud/iss required, allowlist, base64 error handling), CORS (wildcard reject, explicit headers, max_age), JWT strip Bearer, validators (webhook ownership, 400 not 500), IDOR (specific config access), rate limit (sliding window)
- Concurrency: count_message (split phase), roulette modal (move calc outside FOR UPDATE), battlepass claim (atomic), coinsshop (for_update), case add/delete reward (for_update), battlepass add_level (pre-validate), case/color/rainbow locks, valentine cooldown DB-backed, valentine counters locked, compbuilder detached merge fix, clan leave role before delete, role request for_update typo, give/item validation, rainbow batch update before Discord, unpunish dispatch before delete, proposal delete after send, temp_role/multiplier FOR UPDATE SKIP LOCKED + re-fetch, clan leave role before delete
- Security: JWT (Bearer strip, exp/aud/iss required, allowlist), CORS (wildcard reject, explicit headers), webhook SSRF (ownership check), IDOR (specific config access), rate limit (sliding window), permissions (guild_only, deny empty list, UNSAFE audit), webhook ownership
- Tests: test_race_conditions.py (11 checks), ruff E,W,F clean, py_compile passes

+++ src/infra/db/models/configurations/forum.py
@@ -17,7 +17,7 @@ class GuildForumConfig(IdIntegerMixin, Base):
         BigInteger, nullable=False, unique=True
     )
     section_id: Mapped[int | None] = mapped_column(
-        Integer, nullable=True, unique=True
+        BigInteger, nullable=True, unique=True
     )
     _notify_webhook_id: Mapped[int | None] = mapped_column(
         "notify_webhook_id",
